python/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-HTTPAuth==4.0.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
-e git+https://github.com/KoenradusXLVIII/SmartHouse.git@426e563f049c929f520bf9b2dd78f7e501309a72#egg=MQTT&subdirectory=python/MQTT
-e git+https://github.com/KoenradusXLVIII/SmartHouse.git@426e563f049c929f520bf9b2dd78f7e501309a72#egg=P1&subdirectory=python/P1
packaging==20.3
paho-mqtt==1.5.0
phue==1.1
pluggy==0.13.1
psutil==5.7.0
py==1.8.1
PyCRC @ git+https://github.com/alexbutirskiy/PyCRC.git@02b91f44d17dec883f88f998e7847d60180070da
pyparsing==2.4.7
pyserial==3.4
pytest==5.4.1
PyYAML==5.3.1
requests==2.23.0
responses==0.10.14
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
